producer/requirements.txt
kafka-python-ng>=2.2.3
requests==2.31.0